# app/Brand/Resources/BrandResource.php
<?php

namespace App\Brand\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;


/**
 * @OA\Schema(
 *     schema="BrandResource",
 *     title="Brand Resource",
 *     description="Estructura de datos de una marca",
 *     type="object",
 *     @OA\Property(property="id", type="integer", example=1, nullable=false),
 *     @OA\Property(property="name", type="string", example="Nike", nullable=false)
 * )
 */
class BrandResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     */
    public function toArray(Request $request): array
    {
        return [
            'id'        => $this->id,
            'name'      => $this->name
        ];
    }
}